pyproject.toml
[tool.poetry]
name = "kaeri-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
SQLAlchemy = "^2.0.32"
langchain-openai = "^0.1.22"
langchain-core = "^0.2.33"
langchain-milvus = "^0.1.4"
pymilvus = "^2.4.5"
langchain = "^0.2.14"
pydantic-settings = "^2.4.0"
uvicorn = "^0.30.6"
langchain-community = "^0.2.12"
pymupdf = "^1.24.9"
python-multipart = "^0.0.9"
pypdf2 = "^3.0.1"
# redis = "^4.0.0"
redis = "^5.2.1"
openai = ">=1.75.0,<2.0.0"
google-generativeai = "^0.8.5"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.10"
langchain-teddynote = "^0.3.45"
ipython = "^9.4.0"
requests = "^2.32.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
